+++ packages/backend/src/nest/connections-manager/connections-manager.service.ts
@@ -562,6 +562,7 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
       this.serverIoProvider.io.emit(SocketActionTypes.CONNECTION_PROCESS_INFO, data)
     })
     this.storageService.on(StorageEvents.REPLICATED_CERTIFICATES, (payload: SendCertificatesResponse) => {
+      this.logger(`Storage - ${StorageEvents.REPLICATED_CERTIFICATES}`)
       this.serverIoProvider.io.emit(SocketActionTypes.RESPONSE_GET_CERTIFICATES, payload)
     })
     this.storageService.on(StorageEvents.LOAD_PUBLIC_CHANNELS, (payload: ChannelsReplicatedPayload) => {
@@ -628,12 +629,12 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
     )
 
     this.socketService.on(SocketActionTypes.SEND_COMMUNITY_METADATA, async (payload: CommunityMetadata) => {
-      const meta = await this.storageService?.communityMetadataStore?.updateCommunityMetadata(payload)
+      await this.storageService?.communityMetadataStore?.updateCommunityMetadata(payload)
     })
 
     this.storageService.on(StorageEvents.COMMUNITY_METADATA_SAVED, async (meta: CommunityMetadata) => {
       this.logger(`Storage - ${StorageEvents.COMMUNITY_METADATA_SAVED}: ${meta}`)
-      await this.storageService?.certificatesStore?.updateMetadata(meta)
+      this.storageService?.certificatesStore?.updateMetadata(meta)
       this.serverIoProvider.io.emit(SocketActionTypes.COMMUNITY_METADATA_SAVED, meta)
     })
   }