+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 [unreleased]
 
+* Fix: construct all stores before initializing them - initializing community metadata store sets metadata in certificates store
+
 * Add debug logs.
 
 [2.0.3-alpha.14]